src/main/java/com/ucll/Taskmanager/dto/SubTaskDTO.java
package com.ucll.Taskmanager.dto;

import javax.persistence.GeneratedValue;
import javax.persistence.Id;
import javax.validation.constraints.NotEmpty;
import javax.validation.constraints.Size;
import java.util.UUID;

public class SubTaskDTO {

    private UUID id;
    @NotEmpty
    @Size(min=3)
    private String title;
    @NotEmpty
    @Size(max=250)
    private String description;

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }
}
